adapt coko-core tests to OrderDsl changes
